pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "document-genai-assistant"
version = "1.0.0"
description = "A document-aware GenAI assistant for deep document analysis and Q&A"
authors = [
    {name = "GenAI Assistant Team", email = "team@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
keywords = ["ai", "nlp", "document-analysis", "question-answering", "genai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "fastapi>=0.104.1",
    "streamlit>=1.28.1",
    "uvicorn>=0.24.0",
    "google-generativeai>=0.3.2",
    "PyPDF2>=3.0.1",
    "pdfminer.six>=20221105",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.23",
    "langchain>=0.0.335",
    "langchain-community>=0.0.5",
    "langchain-google-genai>=0.0.5",
    "tiktoken>=0.5.1",
    "chromadb>=0.4.15",
    "sentence-transformers>=2.2.2",
    "numpy>=1.24.3",
    "pandas>=2.0.3",
    "pydantic>=2.5.0",
    "requests>=2.31.0",
    "aiofiles>=23.2.0",
    "Pillow>=10.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "mypy>=1.7.1",
]

[project.urls]
Homepage = "https://github.com/example/document-genai-assistant"
Documentation = "https://github.com/example/document-genai-assistant/docs"
Repository = "https://github.com/example/document-genai-assistant"
Issues = "https://github.com/example/document-genai-assistant/issues"

[project.scripts]
genai-assistant = "app.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503", "W504"]
exclude = [
    ".git",
    "__pycache__",
    "dist",
    "build",
    "*.egg",
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --tb=short"
testpaths = [
    "tests",
]
python_files = [
    "test_*.py",
]
python_classes = [
    "Test*",
]
python_functions = [
    "test_*",
]
